submodules: lv_port_linux: Make my fork the URL
In course of initial work on adding OOB experience for LVGL demo on
AM62L, I may make some changes to the sources of the submodules in this
project. As a result, I have forks created for all of these submodules.

Update the URL link for the lv_port_linux submodule to my fork.

Signed-off-by: Suhaas Joshi <s-joshi@ti.com>